convert remaining CI scripts to require UV
