+++ ferrocene/ci/scripts/utils/pyproject.toml
@@ -1,0 +1,11 @@
+# SPDX-License-Identifier: MIT OR Apache-2.0
+# SPDX-FileCopyrightText: The Ferrocene Developers
+
+[project]
+name = "utils"
+version = "0.1.0"
+dependencies = ["boto3", "pyyaml", "zstandard"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
