+++ ferrocene/ci/scripts/utils/src/utils/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: MIT OR Apache-2.0
+# SPDX-FileCopyrightText: The Ferrocene Developers
